gacp

+++ Menu_app/menu2.py
@@ -191,13 +191,13 @@ def menu2(Topics,path):
 def save_topics(Topics,path):
     for customer in Topics['customers']:
         c = get_safe_name(customer)
-        try:
+        if True:#try:
             os.remove(opj(path,'__local__','ready.'+c))
-        except:
+        else:#except:
             pass
-        try:
+        if True:#try:
             os.remove(opj(path,'__local__','Topics.'+c+'.pkl'))
-        except:
+        else:#except:
             pass
         so(Topics,opjh(path,'__local__','Topics.'+c+'.pkl'))
         text_to_file(opjh(path,'__local__','ready.'+c),'')
